backend 2d sprite/GenForgeSprite-main/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-restx>=1.3.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
rembg>=2.0.0
onnxruntime>=1.16.0
